+++ lib/generators/doorkeeper/openid_connect/templates/initializer.rb
@@ -28,6 +28,18 @@ Doorkeeper::OpenidConnect.configure do
     # redirect_to new_user_session_url
   end
 
+  # Depend on the configuration, the DoubleRenderError could be occurred
+  #  because of render/redirect is called at some configuration before this configure is called.
+  # To avoid DoubleRenderError, you could adding these two lines at the beginning
+  #  of this configuration: (Reference: https://github.com/rails/rails/issues/25106)
+  #   self.response_body = nil
+  #   @_response_body = nil
+  select_account_for_resource_owner do |resource_owner, return_to|
+    # Example implementation:
+    # store_location_for resource_owner, return_to
+    # redirect_to account_select_url
+  end
+
   subject do |resource_owner, application|
     # Example implementation:
     # resource_owner.id